finishing the validation task

+++ src/models/sessions.model.js
@@ -19,6 +19,9 @@ exports.getAll = async () => {
 //getBySessionId
 exports.getBySessionId = async (sessionId) => {
 
+    if (typeof sessionId !== 'number')
+        throw new Error("user id (getBySessionId) must be a number");
+
 
     try {
         let res = await dbConnection.query(`SELECT session_id,users.user_id,users.first_name, users.last_name, users.email,users.doctor_certificate 
@@ -35,13 +38,22 @@ exports.getBySessionId = async (sessionId) => {
 
 exports.add = async (userId, duration) => {
 
+    if (typeof userId !== 'number')
+        throw new Error("user id (addSession model) must be a number");
+
+
+    if (typeof duration !== 'number')
+        throw new Error("user id (addSession model) must be a number");
+
+
+
     try {
 
         await dbConnection.query('INSERT INTO sessions (user_id,duration_min) VALUES ($1,$2)', [userId, duration]);
         let res = await dbConnection.query(`SELECT session_id,users.user_id,users.first_name, users.last_name, users.email,users.doctor_certificate 
                                             FROM sessions inner join users on sessions.user_id = users.user_id where sessions.user_id = $1 ; `, [userId]);
 
-        return res.rows[res.rows.length -1];
+        return res.rows[res.rows.length - 1];
 
     } catch (e) {
         throw new Error("An error occured while adding the session to the db")
@@ -53,10 +65,14 @@ exports.add = async (userId, duration) => {
 //delete(SessionID)
 exports.delete = async (sessionId) => {
 
+    if (typeof sessionId !== 'number')
+        throw new Error("user id (delete) must be a number");
+
+
 
     try {
 
-       let res = await dbConnection.query('DELETE FROM sessions WHERE session_id = $1', [sessionId]);
+        let res = await dbConnection.query('DELETE FROM sessions WHERE session_id = $1', [sessionId]);
         return res.rowCount;
 
     } catch (e) {
@@ -68,7 +84,7 @@ exports.delete = async (sessionId) => {
 
 
 //deleteAllExpired()
-exports.deleteAllExpired = async() => {
+exports.deleteAllExpired = async () => {
 
     try {
         let res = await dbConnection.query("DELETE FROM sessions WHERE start_time+(duration_min*INTERVAL'1 minutes')<=now()");
@@ -81,7 +97,11 @@ exports.deleteAllExpired = async() => {
 
 
 //isExpired(SessionID)
-exports.isExpired = async(sessioniId) => {
+exports.isExpired = async (sessioniId) => {
+
+    if (typeof sessionId !== 'number')
+    throw new Error("user id (isExpired - session.model) must be a number");
+
 
     try {
